added check for existing table in TblDelete20171129 to correct broken migration

+++ db/migrate/20171129163200_tbl_delete_20171129.rb
@@ -1,5 +1,5 @@
 class TblDelete20171129 < ActiveRecord::Migration
   def change
-    drop_table :app_configs
+    drop_table :app_configs if table_exists?(:app_configs)
   end
 end
